♻️ refactor: 회원가입 요청 함수 폴더 service -> api로 변경
